Increase upper constraint of puppetlabs-xinetd
The highest version that fits current constraint is 1.5.0, which is > 5
years old[1]. Increase the version so that newer version can be
installed.

The options used for xinetd::service are all present in the newest
version (v3.3.0) of xinetd, so this should be fine.

[1]: https://forge.puppet.com/modules/puppetlabs/xinetd/1.5.0

Change-Id: Id31370dbc69361f6857e9c9825d7fe1c430954e0
(cherry picked from commit eff29bb202aa4c019a10f0882e13cb1e77caec9e)